requirements.txt
requests==2.9.1
paho-mqtt==1.1
APScheduler==3.0.5
